+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Versions
 
+## 1.6.6
+
+- Retry polling peers for L-BTC balances after peerswap initializes
+
 ## 1.6.5
 
-- Reduce frequency of balance announcements to 24hr unless changed
+- Reduce frequency of balance announcements to daily unless changed
 - Pre-fill 0 if possible swap amount is below 100,000
 
 ## 1.6.4